@@ -1,6 +1,6 @@
 {
     "name": "padhie/twitch-api-bundle",
-    "version": "1.1.15",
+    "version": "1.2.0",
     "type": "library",
     "keywords": [
         "twitch",
@@ -18,7 +18,7 @@
         }
     ],
     "require": {
-        "php": "^7.1 || ^8",
+        "php": "^7.4 || ^8",
         "ext-curl": "*",
         "ext-json": "*",
         "guzzlehttp/guzzle": "^6.3"
@@ -39,6 +39,6 @@
     },
     "scripts": {
         "phpunit": ["vendor/bin/phpunit tests"],
-        "phpstan": ["vendor/bin/phpstan analyse -l 7 --ansi src"]
+        "phpstan": ["vendor/bin/phpstan analyse -c phpstan.neon --ansi"]
     }
 }
